Rain {CRF}	R Documentation

Rain data

Description

This data set gives an example of rain data used to train CRF and MRF models

Usage

data(Rain)

Format

A list containing two elements:

rain A matrix of 28 columns containing raining data (1: rain, 2: sunny). Each row is an instance of 28 days for one month.

months A vector containing the months of each instance.

References

Mark Schmidt. UGM: Matlab code for undirected graphical models. http://www.di.ens.fr/~mschmidt/Software/UGM.html

[Package CRF version 0.3-10 Index]