decode {CRF}	R Documentation

Decoding methods

Description

Computing the most likely configuration

Usage


decode.exact(crf)
decode.chain(crf)
decode.tree(crf)
decode.conditional(crf, clamped, decode.method, ...)
decode.cutset(crf, cutset, engine = "default",
              start = apply(crf$node.pot, 1, which.max))
decode.junction(crf)
decode.sample(crf, sample.method, ...)
decode.marginal(crf, infer.method, ...)
decode.lbp(crf, max.iter = 10000, cutoff = 1e-4, verbose = 0)
decode.trbp(crf, max.iter = 10000, cutoff = 1e-4, verbose = 0)
decode.greedy(crf, restart = 0, start = apply(crf$node.pot, 1, which.max))
decode.icm(crf, restart = 0, start = apply(crf$node.pot, 1, which.max))
decode.block(crf, blocks, decode.method = decode.tree, restart = 0,
             start = apply(crf$node.pot, 1, which.max), ...)
decode.ilp(crf, lp.rounding = FALSE)

Arguments

crf

clamped

cutset

decode.method

infer.method

sample.method

engine

start

max.iter

cutoff

verbose

restart

blocks

lp.rounding

...

Details

decode.exact Exact decoding for small graphs with brute-force search

decode.chain Exact decoding for chain-structured graphs with the Viterbi algorithm

decode.tree Exact decoding for tree- and forest-structured graphs with max-product belief propagation

decode.conditional Conditional decoding (takes another decoding method as input)

decode.cutset Exact decoding for graphs with a small cutset using cutset conditioning

decode.junction Exact decoding for low-treewidth graphs using junction trees

decode.sample Approximate decoding using sampling (takes a sampling method as input)

decode.marginal Approximate decoding using inference (takes an inference method as input)

decode.lbp Approximate decoding using max-product loopy belief propagation

decode.trbp Approximate decoding using max-product tree-reweighted belief propagtion

decode.greedy Approximate decoding with greedy algorithm

decode.icm Approximate decoding with the iterated conditional modes algorithm

decode.block Approximate decoding with the block iterated conditional modes algorithm

decode.ilp Exact decoding with an integer linear programming formulation and approximate using LP relaxation

See Also

make.crf, infer, sample

Examples


library(CRF)
data(Small)
d <- decode.exact(small.crf)
d <- decode.chain(small.crf)
d <- decode.tree(small.crf)
d <- decode.conditional(small.crf, c(0,1,0,0), decode.exact)
d <- decode.cutset(small.crf, c(2))
d <- decode.junction(small.crf)
d <- decode.sample(small.crf, sample.exact, 10000)
d <- decode.marginal(small.crf, infer.exact)
d <- decode.lbp(small.crf)
d <- decode.trbp(small.crf)
d <- decode.greedy(small.crf)
d <- decode.icm(small.crf)
d <- decode.block(small.crf, list(c(1,3), c(2,4)))
d <- decode.ilp(small.crf)

[Package CRF version 0.2-4 Index]