decode.cutset {CRF}	R Documentation

Decoding method for graphs with a small cutset

Description

Computing the most likely configuration for CRF

Usage

decode.cutset(crf, cutset, engine = "default", start = apply(crf$node.pot,
  1, which.max))

Arguments

crf
The CRF

cutset
A vector of nodes in the cutset

engine
The underlying engine for cutset decoding, possible values are "default", "none", "exact", "chain", and "tree".

start
An initial configuration, a good start will significantly reduce the seraching time

Details

Exact decoding for graphs with a small cutset using cutset conditioning

Value

This function will return the most likely configuration, which is a vector of length crf$n.nodes.

Examples

library(CRF)
data(Small)
d <- decode.cutset(Small$crf, c(2))

[Package CRF version 0.3-5 Index]