decode.greedy {CRF}	R Documentation

Decoding method using greedy algorithm

Description

Computing the most likely configuration for CRF

Usage

decode.greedy(crf, restart = 0, start = apply(crf$node.pot, 1, which.max))

Arguments

crf
The CRF

restart
Non-negative integer to control how many restart iterations are repeated

start
An initial configuration, a good start will significantly reduce the seraching time

Details

Approximate decoding with greedy algorithm

Value

This function will return the most likely configuration, which is a vector of length crf$n.nodes.

Examples

library(CRF)
data(Small)
d <- decode.greedy(Small$crf)

[Package CRF version 0.3-10 Index]