decode.icm {CRF}	R Documentation

Decoding method using iterated conditional modes algorithm

Description

Computing the most likely configuration for CRF

Usage

decode.icm(crf, restart = 0, start = apply(crf$node.pot, 1, which.max))

Arguments

crf
The CRF

restart
Non-negative integer to control how many restart iterations are repeated

start
An initial configuration, a good start will significantly reduce the seraching time

Details

Approximate decoding with the iterated conditional modes algorithm

Value

This function will return the most likely configuration, which is a vector of length crf$n.nodes.

Examples

library(CRF)
data(Small)
d <- decode.icm(Small$crf)

[Package CRF version 0.3-8 Index]