decode.ilp {CRF}	R Documentation

Decoding method using integer linear programming

Description

Computing the most likely configuration for CRF

Usage

decode.ilp(crf, lp.rounding = FALSE)

Arguments

crf
The CRF

lp.rounding
Boolean variable to indicate whether LP rounding is need.

Details

Exact decoding with an integer linear programming formulation and approximate using LP relaxation

Value

This function will return the most likely configuration, which is a vector of length crf$n.nodes.

Examples

library(CRF)
data(Small)
d <- decode.ilp(Small$crf)

[Package CRF version 0.3-8 Index]