decode.lbp {CRF}	R Documentation

Decoding method using loopy belief propagation

Description

Computing the most likely configuration for CRF

Usage

decode.lbp(crf, max.iter = 10000, cutoff = 1e-04, verbose = 0)

Arguments

crf
The CRF

max.iter
The maximum allowed iterations of termination criteria

cutoff
The convergence cutoff of termination criteria

verbose
Non-negative integer to control the tracing informtion in algorithm

Details

Approximate decoding using max-product loopy belief propagation

Value

This function will return the most likely configuration, which is a vector of length crf$n.nodes.

Examples

library(CRF)
data(Small)
d <- decode.lbp(Small$crf)

[Package CRF version 0.3-9 Index]