duplicate.crf {CRF}	R Documentation

Duplicate CRF

Description

Duplicate an existing CRF

Usage

duplicate.crf(crf)

Arguments

crf
The existing CRF

Details

This function will duplicate an existing CRF. Since CRF is implemented as an environment, normal assignment will only copy the pointer instead of the real data. This function will generate a new CRF and really copy all data.

Value

The function will return a new CRF with copied data

See Also

make.crf

[Package CRF version 0.3-7 Index]