gradient {CRF}	R Documentation

Calculate CRF negative log-likelihood gradient

Description

Calculate the gradient of negative log likelihood of CRF model

Usage

gradient(par, crf, ...)

Arguments

par

crf

...

Details

Calculate the gradient of negative log likelihood of CRF model. This function is used by optimization algorithm in training.

Value

This function will return the gradient of CRF negative log-likelihood.

[Package CRF version 0.3-4 Index]