make.par {CRF}	R Documentation

Make CRF parameters

Description

Make the data structure of CRF parameters

Usage

make.par(crf, n.par = 1)

Arguments

crf
The CRF

n.par
The number of parameters

Details

This function makes the data structure of parameters need for modeling and training CRF. The parameters are stored in par, which is a numeric vector of length n.par.

Value

This function will directly modify the CRF and return the same CRF.

See Also

crf.update, make.features, make.crf

[Package CRF version 0.3-10 Index]