mrf.stat {CRF}	R Documentation

Calculate MRF sufficient statistics

Description

Calculate the sufficient statistics of MRF model

Usage

mrf.stat(crf, instances)

Arguments

crf
The CRF

instances
The training data matrix of MRF model

Details

This function calculates the sufficient statistics of MRF model. This function much be called before the first calling to mrf.nll. In the training data matrix instances, each row is an instance and each column corresponds a node in CRF.

Value

This function will return the value of MRF sufficient statistics.

See Also

mrf.nll, train.mrf

[Package CRF version 0.3-6 Index]