sample.chain {CRF}	R Documentation

Sampling method for chain-structured graphs

Description

Generating samples from the distribution

Usage

sample.chain(crf, size)

Arguments

crf
The CRF

size
The sample size

Details

Exact sampling for chain-structured graphs with the forward-filter backward-sample algorithm

Value

This function will return a matrix with size rows and crf$n.nodes columns, in which each row is a sampled configuration.

Examples

library(CRF)
data(Small)
s <- sample.chain(Small$crf, 100)

[Package CRF version 0.3-6 Index]