sample.exact {CRF}	R Documentation

Sampling method for small graphs

Description

Generating samples from the distribution

Usage

sample.exact(crf, size)

Arguments

crf

size

Details

Exact sampling for small graphs with brute-force inverse cumulative distribution

Value

This function will return a matrix with size rows and crf$n.nodes columns, in which each row is a sampled configuration.

Examples

library(CRF)
data(Small)
s <- sample.exact(small.crf, 100)

[Package CRF version 0.3-4 Index]