sample.gibbs {CRF}	R Documentation

Sampling method using single-site Gibbs sampler

Description

Generating samples from the distribution

Usage

sample.gibbs(crf, size, burn.in = 1000, start = apply(crf$node.pot, 1,
  which.max))

Arguments

crf

size

burn.in

start

Details

Approximate sampling using a single-site Gibbs sampler

Value

This function will return a matrix with size rows and crf$n.nodes columns, in which each row is a sampled configuration.

Examples

library(CRF)
data(Small)
s <- sample.gibbs(small.crf, 100)

[Package CRF version 0.3-4 Index]