sub.crf {CRF}	R Documentation

Make sub CRF data structure

Description

Generate sub CRF data structure by selecting some nodes

Usage


clamp.crf(crf, clamped)

Arguments

crf
The CRF data structure generated by make.crf

subset
The vector of selected node ids

Details

The function will generate a CRF data structure from a given CRF data structure by selecting some nodes. The vector subset contains the node ids selected to generate the new CRF data. Unlike clamp.crf, the potentials of remainning nodes and edges are untouched.

Value

The function will return a CRF data structure with additional components:

original
The original CRF data.

node.id
The vector of the original node ids for nodes in the new CRF data.

node.map
The vector of the new node ids for nodes in the original CRF data.

edge.id
The vector of the original edge ids for edges in the new CRF data.

edge.map
The vector of the new edge ids for edges in the original CRF data.

See Also

make.crf, clamp.crf

Examples


library(CRF)
data(Small)
crf <- sub.crf(small.crf, c(2, 3))


[Package CRF version 0.1-6 Index]