train.mrf {CRF}	R Documentation

Train MRF model

Description

Train the MRF model to estimate the parameters

Usage

train.mrf(crf, instances, nll = mrf.nll, trace = 0)

Arguments

crf
The CRF

instances
The training data matrix of CRF model

nll
The function to calculate negative log likelihood

trace
Non-negative integer to control the tracing informtion of the optimization process

Details

This function trains the Markov Random Fields (MRF) model, which is a simple variant of CRF model.

In the training data matrix instances, each row is an instance and each column corresponds a node in CRF.

Value

This function will directly modify the CRF and return the same CRF.

See Also

mrf.update, mrf.stat, mrf.nll, make.crf

[Package CRF version 0.3-8 Index]